remove support for page 1
